refactor(core): 시각화 코드 분리 — HRIRPlotter / ImpulseResponsePlotter (이슈 #87 Phase 1)
`core/plotting/` 서브패키지를 신설해 matplotlib/Bokeh 시각화 로직을 데이터
클래스에서 분리한다. `HRIR(HRIRPlotter)`, `ImpulseResponse(ImpulseResponsePlotter)`
mixin 상속으로 기존 호출 API(`hrir.plot(...)`, `ir.plot_fr(...)` 등)는 그대로 유지.

- core/hrir.py: 2086 → 1149 라인 (~937 라인 감소)
- core/impulse_response.py: 1231 → 534 라인 (~697 라인 감소)
- 두 파일 모두 이슈 #87 목표인 1,500 라인 미만 달성
- BRIR md5는 기본값/`--vbass` 두 경로 모두 master와 일치(Windows 검증):
  - 기본값: cf37a9aaf95e717c04e309fba05fa61d
  - vbass: 07eef9ef89cc0d55313c9c0c18edbc76
